refactor: move content and toolbar to sepetate component
